requirements.txt
asyncio
websockets
pydantic
jsonschema
timed_rotating_text_file
fastapi
fastapi_pagination
uvicorn
psycopg2-binary
sqlalchemy
asyncpg
fastapi_sqla